Update README.md
The LMS is a console-based application developed in Java. It lets users, such as librarian or their staff members, to manage a library's collection of books. Management includes adding and removing books for a library's book collection. Input/ Output: Reads book information from a text file and saves changes back to file. OOP: The project utilizes object-oriented programming by defining classes such as "Book" and "Library". Includes exception handling where the file is not found during loading or it there's an issue while writing to the file.